+++ src/const.js
@@ -37,6 +37,15 @@ const POINT_MAX_PRICE = 1000;
 const POINTS__MAX_COUNT = 10;
 const PICTURE_MAX = 5;
 
+const NO_POINT_MASSAGES = {
+  loadFail: 'Failed to load latest route information',
+  loading: 'Loading...',
+  everthing: 'Click New Event to create your first point',
+  past: 'There are no past events now',
+  present: 'There are no present events now',
+  future: 'There are no future events now'
+};
+
 export {
   POINT_TYPES,
   DESTINATIONS,
@@ -47,7 +56,8 @@ export {
   MILISECONDS,
   POINT_MAX_PRICE,
   POINTS__MAX_COUNT,
-  PICTURE_MAX
+  PICTURE_MAX,
+  NO_POINT_MASSAGES
 };
