+++ src/mock/destinations-mock.js
@@ -1,4 +1,4 @@
-import { getRandomInteger } from '../utils.js';
+import { getRandomInteger } from '../utils/common.js';
 import { DESCRIPTIONS, DESTINATIONS, PICTURE_MAX } from '../const.js';
 
 const createDestinationUrl = (id) => (
